Make RemoteGroup and RemoteLight extend RemoteEndpointEntity.
